docker-storage-setup: Fix typo and condition issues
If 'VG_SIZE' is empty then will hit a shell syntax issue, so it has better to
make sure the 'VG_SIZE' is non-null firstly. In addition, option '-L' of
lvcreate requires an argument to create a logical volume, in here, it
should make sure the 'META_SIZE' not 'META_LV_SIZE' is non-null too.

Signed-off-by: Alex Jia <ajia@redhat.com>