[WIP]: Don't do anything if Docker already has data on loopback
(This patch is not tested, submitting to gather comments)

If we want this service to start by default, it doesn't make sense to
try to allocate thinpool space if there are already Docker images on
loopback.  People would then just see their images/containers vanish.

Let's not do anything unless the system appears to be a clean slate.

For example, we also now silently do nothing if the user has
explicitly set a backend with `-s`, such as `-s overlayfs`.