Introduce `auto` driver
This currently checks whether we have `ftype=1` on XFS; down the line I might
enhance this for RHEL to detect a new enough kernel. Prep for switching RHEL
Atomic Host in 7.5 to overlay by default.

I changed the code to consistently use a `case` statement for driver matching,
pretending a bit like the driver list is a C `enum` and we're doing a `switch`.